Add a cancel button for new game dialog, Refactor move files, Cleanup
