+++ lib/common/presentation/clean_players_view_model.dart
@@ -1,6 +1,6 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
-import 'package:tagros_comptes/state/providers.dart';
 import 'package:tagros_comptes/tagros/data/source/db/app_database.dart';
+import 'package:tagros_comptes/tagros/data/source/db/db_providers.dart';
 
 class CleanPlayer extends AutoDisposeNotifier<Stream<List<Player>>> {
   @override
