+++ lib/tagros/presentation/dialog/dialog_players.dart
@@ -3,8 +3,8 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:tagros_comptes/common/presentation/component/player_search_field.dart';
 import 'package:tagros_comptes/common/presentation/navigation.dart';
 import 'package:tagros_comptes/generated/l10n.dart';
-import 'package:tagros_comptes/state/viewmodel/choose_player_view_model.dart';
 import 'package:tagros_comptes/tagros/data/game_provider.dart';
+import 'package:tagros_comptes/tagros/presentation/dialog/choose_player_view_model.dart';
 
 class DialogChoosePlayers extends ConsumerWidget {
   const DialogChoosePlayers({super.key});
@@ -24,6 +24,11 @@ class DialogChoosePlayers extends ConsumerWidget {
             return const SizedBox();
           }
         })),
+        OutlinedButton(
+            onPressed: () {
+              Navigator.of(context).pop();
+            },
+            child: Text(S.of(context).cancel)),
         ElevatedButton(
             onPressed: () {
               if (ref.read(choosePlayerProvider.notifier).validatePlayers()) {
@@ -34,7 +39,7 @@ class DialogChoosePlayers extends ConsumerWidget {
                         ref.read(choosePlayerProvider).$2));
               }
             },
-            child: Text(S.of(context).ok))
+            child: Text(S.of(context).ok)),
       ],
     );
   }
